third_party/WebKit/LayoutTests/fast/events/focus-event-source-device-from-keyboard-expected.txt


Test that focus events from keyboard taps should have a sourceDevice and its firesTouchEvents should be false.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


tests on div
focus
PASS event.sourceDevice is non-null.
PASS event.sourceDevice.firesTouchEvents is false
focusin
PASS event.sourceDevice is non-null.
PASS event.sourceDevice.firesTouchEvents is false

tests on input
blur
PASS event.sourceDevice is non-null.
PASS event.sourceDevice.firesTouchEvents is false
focusout
PASS event.sourceDevice is non-null.
PASS event.sourceDevice.firesTouchEvents is false
focus
PASS event.sourceDevice is non-null.
PASS event.sourceDevice.firesTouchEvents is false
focusin
PASS event.sourceDevice is non-null.
PASS event.sourceDevice.firesTouchEvents is false

tests on button
blur
PASS event.sourceDevice is non-null.
PASS event.sourceDevice.firesTouchEvents is false
focusout
PASS event.sourceDevice is non-null.
PASS event.sourceDevice.firesTouchEvents is false
focus
PASS event.sourceDevice is non-null.
PASS event.sourceDevice.firesTouchEvents is false
focusin
PASS event.sourceDevice is non-null.
PASS event.sourceDevice.firesTouchEvents is false

tests on textarea
blur
PASS event.sourceDevice is non-null.
PASS event.sourceDevice.firesTouchEvents is false
focusout
PASS event.sourceDevice is non-null.
PASS event.sourceDevice.firesTouchEvents is false
focus
PASS event.sourceDevice is non-null.
PASS event.sourceDevice.firesTouchEvents is false
focusin
PASS event.sourceDevice is non-null.
PASS event.sourceDevice.firesTouchEvents is false

tests on select
blur
PASS event.sourceDevice is non-null.
PASS event.sourceDevice.firesTouchEvents is false
focusout
PASS event.sourceDevice is non-null.
PASS event.sourceDevice.firesTouchEvents is false
focus
PASS event.sourceDevice is non-null.
PASS event.sourceDevice.firesTouchEvents is false
focusin
PASS event.sourceDevice is non-null.
PASS event.sourceDevice.firesTouchEvents is false

tests on a
blur
PASS event.sourceDevice is non-null.
PASS event.sourceDevice.firesTouchEvents is false
focusout
PASS event.sourceDevice is non-null.
PASS event.sourceDevice.firesTouchEvents is false
focus
PASS event.sourceDevice is non-null.
PASS event.sourceDevice.firesTouchEvents is false
focusin
PASS event.sourceDevice is non-null.
PASS event.sourceDevice.firesTouchEvents is false

PASS successfullyParsed is true

TEST COMPLETE
